AULA04/ex4/EdgeSolution/modules/ServoModulegsb/requirements.txt
azure-iot-device~=2.7.0
RPi.GPIO